FTP New development
git-svn-id: http://hobbs.evolvus.com/source-repository/FTP_KUWAIT/trunk@42 733bd532-2ea2-478c-8357-b8447c0750e9

+++ ftp-kuwait/src/main/java/com/evolvus/abk/ftp/domain/rates/CurrencyRates.java
@@ -42,10 +42,10 @@ public class CurrencyRates {
 	private String tenor;
 	
 	@Column(name="DAYS_FROM")
-	private Double daysFrom;
+	private Integer daysFrom;
 	
 	@Column(name="DAYS_TO")
-	private Double daysTo;
+	private Integer daysTo;
 	
 	@Column(name="BASE", columnDefinition="decimal(17,7)")
 	private BigDecimal base;
